Fixed plow vector obstacle search radius
Vehicle is striking known obstacles. The routine to generate plow-vectors was only looking at the immediate cell that the plow would pass through for obstructions and not searching the collision radius around it. Plow vector routine is now using the collision radius margin of safety when generating plow-vectors. [Not live tested]